Presentation Deployment

+++ src/main/java/com/cs203/locus/controllers/EventTicketController.java
@@ -1,45 +1,56 @@
 package com.cs203.locus.controllers;
 
-import java.util.ArrayList;
-
 import com.cs203.locus.models.event.Event;
 import com.cs203.locus.models.event.EventTicket;
 import com.cs203.locus.models.event.EventTicketDTO;
 import com.cs203.locus.models.participant.Participant;
 import com.cs203.locus.service.EventService;
 import com.cs203.locus.service.EventTicketService;
-
 import com.cs203.locus.service.ParticipantService;
+import com.cs203.locus.util.EmailUtil;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import com.cs203.locus.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.parameters.P;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
-//TODO: Remaining PUT method
+
 @RestController
 @RequestMapping("/ticket")
 public class EventTicketController {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(EventTicketController.class);
+
     @Autowired
     private EventTicketService eventTicketService;
-
     @Autowired
     private EventService eventService;
-
     @Autowired
     private ParticipantService participantService;
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private EmailUtil emailUtil;
 
 
-    // TODO: is this needed?
-    @GetMapping(value = "/list")
-    public @ResponseBody ResponseEntity<?> getAllEventTickets() {
-        Iterable<EventTicket> temp = eventTicketService.findAll();
+    @GetMapping(value = "/list/{id}")
+    public @ResponseBody
+    ResponseEntity<?> getAllEventTicketsByEventID(@PathVariable Integer id) {
+        Iterable<EventTicket> temp = eventTicketService.findEventTicketByEventId(id);
         ArrayList<EventTicketDTO> result = new ArrayList<>();
         for (EventTicket eventTicket : temp) {
             EventTicketDTO toRet = new EventTicketDTO();
+            toRet.setIsVaccinated(eventTicket.getParticipant().getVaxStatus());
             toRet.setId(eventTicket.getId());
             toRet.setParticipantName(eventTicket.getParticipant().getUser().getName());
             toRet.setParticipantId(eventTicket.getParticipant().getId());
@@ -55,9 +66,10 @@ public class EventTicketController {
 
         return ResponseEntity.ok(result);
     }
-    
+
     @GetMapping("/{id}")
-    public @ResponseBody ResponseEntity<EventTicketDTO> findById(@PathVariable Integer id){
+    public @ResponseBody
+    ResponseEntity<EventTicketDTO> findById(@PathVariable Integer id) {
         EventTicket result = eventTicketService.findById(id);
 
         if (result == null) {
@@ -66,8 +78,12 @@ public class EventTicketController {
         }
 
         EventTicketDTO toRet = new EventTicketDTO();
+        toRet.setId(result.getId());
         toRet.setParticipantName(result.getParticipant().getUser().getName());
+        toRet.setParticipantId(result.getParticipant().getId());
+        toRet.setIsVaccinated(result.getParticipant().getVaxStatus());
         toRet.setOrganiserName(result.getEvent().getOrganiser().getUser().getName());
+        toRet.setOrganiserId(result.getEvent().getOrganiser().getId());
         toRet.setEventName(result.getEvent().getName());
         toRet.setEventId(result.getEvent().getId());
         toRet.setStartDateTime(result.getEvent().getStartDateTime());
@@ -78,8 +94,9 @@ public class EventTicketController {
     }
 
     // TODO: ensure participant can only access his own list of EventTickets
-    @GetMapping(value ="/listParticipantTickets/{id}")
-    public @ResponseBody ResponseEntity<ArrayList<EventTicketDTO>> getParticipantTickets(@PathVariable Integer id){
+    @GetMapping(value = "/listParticipantTickets/{id}")
+    public @ResponseBody
+    ResponseEntity<ArrayList<EventTicketDTO>> getParticipantTickets(@PathVariable Integer id) {
         if (participantService.findById(id) == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
                     "No Participant with ID: " + id);
@@ -90,6 +107,7 @@ public class EventTicketController {
         for (EventTicket eventTicket : temp) {
             EventTicketDTO toRet = new EventTicketDTO();
             toRet.setId(eventTicket.getId());
+            toRet.setIsVaccinated(eventTicket.getParticipant().getVaxStatus());
             toRet.setParticipantName(eventTicket.getParticipant().getUser().getName());
             toRet.setParticipantId(eventTicket.getParticipant().getId());
             toRet.setOrganiserName(eventTicket.getEvent().getOrganiser().getUser().getName());
@@ -105,9 +123,15 @@ public class EventTicketController {
         return ResponseEntity.ok(result);
     }
 
+    @GetMapping(value = "/hasParticipatedEvent/{participantId}/{eventId}")
+    ResponseEntity<Boolean> getParticpationStatus(@PathVariable Integer participantId, @PathVariable Integer eventId){
+        return ResponseEntity.ok(eventTicketService.existingTicket(participantId,eventId));
+    }
+
     // Identify event ticket exists using eventId and userId
-    @GetMapping(value ="/listParticipantTickets/{id}/{eventId}")
-    public @ResponseBody ResponseEntity<ArrayList<EventTicketDTO>> getParticipantTickets(@PathVariable Integer id, @PathVariable Integer eventId){
+    @GetMapping(value = "/listParticipantTickets/{id}/{eventId}")
+    public @ResponseBody
+    ResponseEntity<ArrayList<EventTicketDTO>> getParticipantTickets(@PathVariable Integer id, @PathVariable Integer eventId) {
         if (eventTicketService.findSpecificTicket(id, eventId) == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
                     "No Ticket with Id: " + id + " and eventId: " + eventId);
@@ -118,6 +142,7 @@ public class EventTicketController {
         for (EventTicket eventTicket : temp) {
             EventTicketDTO toRet = new EventTicketDTO();
             toRet.setId(eventTicket.getId());
+            toRet.setIsVaccinated(eventTicket.getParticipant().getVaxStatus());
             toRet.setParticipantName(eventTicket.getParticipant().getUser().getName());
             toRet.setParticipantId(eventTicket.getParticipant().getId());
             toRet.setOrganiserName(eventTicket.getEvent().getOrganiser().getUser().getName());
@@ -133,15 +158,28 @@ public class EventTicketController {
         return ResponseEntity.ok(result);
     }
 
-    // TODO: ensure only participant can create an EventTicket for himself
     @PostMapping("/new")
+    public @ResponseBody
+    ResponseEntity<EventTicketDTO> addTicket(@RequestParam Integer participantId,
+                                             @RequestParam(required = false) Integer eventId,
+                                             @RequestParam(required = false) String inviteCode) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (!userService.findByUsername(auth.getName()).getId().equals(participantId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
 
-    public @ResponseBody ResponseEntity<EventTicketDTO> addTicket(@RequestParam Integer participantId, @RequestParam Integer eventId) {
+        Event event;
+        if (eventId == null) {
+            event = eventService.findByInviteCode(inviteCode);
+        } else if (inviteCode == null) {
+            event = eventService.findById(eventId);
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
 
-        Event event = eventService.findById(eventId);
         if (event == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
-                    "No Event with ID: " + eventId);
+                    "Invalid event");
         }
 
         Participant participant = participantService.findById(participantId);
@@ -160,6 +198,7 @@ public class EventTicketController {
         EventTicket created = eventTicketService.addTicket(ticket);
         EventTicketDTO toRet = new EventTicketDTO();
         toRet.setId(created.getId());
+        toRet.setIsVaccinated(created.getParticipant().getVaxStatus());
         toRet.setParticipantName(created.getParticipant().getUser().getName());
         toRet.setParticipantId(created.getParticipant().getId());
         toRet.setOrganiserName(created.getEvent().getOrganiser().getUser().getName());
@@ -170,12 +209,36 @@ public class EventTicketController {
         toRet.setEndDateTime(created.getEvent().getEndDateTime());
         toRet.setEventAddress(created.getEvent().getAddress());
 
+
+        // Send the Email
+        Map<String, Object> formModel = new HashMap<>();
+        formModel.put("recipientEmailAddress", created.getParticipant().getUser().getEmail());
+        formModel.put("userName", created.getParticipant().getUser().getName());
+        formModel.put("eventName", created.getEvent().getName());
+        formModel.put("eventId", created.getEvent().getId());
+
+        // Send an Email to the organiser to let them know they have successfully created the event
+        try {
+            emailUtil.sendEventSignUpEmail(formModel);
+        }catch (Exception e){
+            LOGGER.error(e.getMessage());
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,
+                    "Unknown error occurs, please try again!");
+        }
+      
         return ResponseEntity.ok(toRet);
     }
 
-    // TODO: ensure only participant can delete own EventTicket
     @DeleteMapping("/{id}")
-    public @ResponseBody ResponseEntity<EventTicket> deleteWithId(@PathVariable Integer id) {
+    public @ResponseBody
+    ResponseEntity<EventTicket> deleteWithId(@PathVariable Integer id) {
+        EventTicket toDel = eventTicketService.findById(id);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        Integer authParticipantId = userService.findByUsername(username).getId();
+        if (!toDel.getParticipant().getId().equals(authParticipantId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         EventTicket result = eventTicketService.deleteById(id);
         if (result == null) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
